optimize group page
