+++ web/src/interfaces/Balances.ts
@@ -1,0 +1,5 @@
+import { PersonFinancials } from "./PersonFinancials";
+
+export interface Balances {
+  [personId: string]: PersonFinancials;
+}
