+++ web/src/pages/Group.tsx
@@ -16,28 +16,16 @@ import { DebtorsExpense } from "../interfaces/DebtorsExpense";
 import { Expense } from "../interfaces/Expense";
 import { Group as GroupType } from "../interfaces/Group";
 import PersonInfo from "../components/PersonInfo";
-import {
-  Card,
-  Heading,
-  Text,
-  Container,
-  VStack,
-  HStack,
-  Button,
-  Input,
-  Box,
-  Badge,
-  SimpleGrid,
-  Stat,
-  StatLabel,
-  StatNumber,
-  StatHelpText,
-  IconButton,
-} from "@chakra-ui/react";
-import { DeleteIcon } from "@chakra-ui/icons";
+import { Text, Container, VStack, Button, Box } from "@chakra-ui/react";
 import { useNotifications } from "../hooks/useNotifications";
 import { ExpenseCreateWithoutGroupId } from "../interfaces/ExpenseCreate";
-import { PersonFinancials } from "../interfaces/PersonFinancials";
+import AddPersonSection from "../components/AddPersonSection";
+import PersonsList from "../components/PersonsSection";
+import ExpensesList from "../components/ExpensesList";
+import BalancesSummary from "../components/BalancesSummary";
+import GroupHeader from "../components/GroupHeader";
+import GroupNotFound from "../components/GroupNotFound";
+import { Balances } from "../interfaces/Balances";
 
 const Group = () => {
   const { groupId } = useParams();
@@ -46,14 +34,8 @@ const Group = () => {
   const [persons, setPersons] = useState<Person[]>([]);
   const [expenses, setExpenses] = useState<Expense[]>([]);
   const [debtorsExpenses, setDebtorsExpenses] = useState<DebtorsExpense[]>([]);
-  const [balances, setBalances] = useState<{
-    [personId: string]: PersonFinancials;
-  } | null>(null);
+  const [balances, setBalances] = useState<Balances | null>(null);
   const [isLoading, setIsLoading] = useState(false);
-
-  // Add person state
-  const [newPersonName, setNewPersonName] = useState("");
-  const [isAddingPerson, setIsAddingPerson] = useState(false);
 
   // Expense popup state
   const [isPopupOpen, setIsPopupOpen] = useState(false);
@@ -65,20 +47,15 @@ const Group = () => {
 
   const { showError, showSuccess } = useNotifications();
 
-  const handleAddPerson = async () => {
-    if (!newPersonName.trim() || !groupId) {
-      showError({
-        title: "Error",
-        description: "Please enter a person's name",
-      });
-      return;
-    }
+  const handleAddExpense = (clickedPersonId: string) => {
+    setIsPopupOpen(true);
+    setClickedPersonId(clickedPersonId);
+  };
 
-    setIsAddingPerson(true);
+  const handlePersonsChange = async (newPersonName: string) => {
     try {
-      const { person: newPerson } = await addPerson(newPersonName, groupId);
+      const { person: newPerson } = await addPerson(newPersonName, groupId!);
       setPersons([...persons, newPerson[0]]);
-      setNewPersonName("");
       showSuccess({
         title: "Success",
         description: `${newPerson.name} added to the group!`,
@@ -90,14 +67,7 @@ const Group = () => {
         title: "Error",
         description: "Failed to add person. Please try again.",
       });
-    } finally {
-      setIsAddingPerson(false);
     }
-  };
-
-  const handleAddExpense = (clickedPersonId: string) => {
-    setIsPopupOpen(true);
-    setClickedPersonId(clickedPersonId);
   };
 
   const handleSubmitExpense = async (data: ExpenseCreateWithoutGroupId) => {
@@ -228,189 +198,46 @@ const Group = () => {
     );
   }
 
-  if (!group) {
-    return (
-      <Container maxW="6xl" py={8}>
-        <VStack spacing={4}>
-          <Text textAlign="center" color="red.500">
-            Group not found. Please check the group ID.
-          </Text>
-          <Link to="/">
-            <Button colorScheme="blue">Go Home</Button>
-          </Link>
-        </VStack>
-      </Container>
-    );
+  if (!group || !groupId) {
+    return <GroupNotFound />;
   }
 
   return (
     <Container maxW="6xl" py={8}>
       <VStack spacing={6} align="stretch">
         {/* Header */}
-        <Box textAlign="center">
-          <Heading size="xl" mb={2}>
-            {group.name}
-          </Heading>
-          <Text color="gray.500" fontSize="sm">
-            Group ID: {groupId}
-          </Text>
-          <Text color="gray.600" fontSize="sm" mt={1}>
-            Share this URL with others to let them join and add expenses
-          </Text>
-        </Box>
+        <GroupHeader groupName={group.name} groupId={groupId!} />
 
         {/* Add Person Section */}
-        <Card p={4}>
-          <VStack spacing={3}>
-            <Heading size="md">Add Person to Group</Heading>
-            <HStack w="100%">
-              <Input
-                placeholder="Enter person's name"
-                value={newPersonName}
-                onChange={(e) => setNewPersonName(e.target.value)}
-                onKeyDown={(e) => e.key === "Enter" && handleAddPerson()}
-              />
-              <Button
-                colorScheme="green"
-                onClick={handleAddPerson}
-                isLoading={isAddingPerson}
-                loadingText="Adding..."
-              >
-                Add Person
-              </Button>
-            </HStack>
-          </VStack>
-        </Card>
+        <AddPersonSection
+          groupId={groupId}
+          onPersonsChange={handlePersonsChange}
+        />
 
         {/* Persons Section */}
-        <Box>
-          <Heading size="lg" mb={4}>
-            People ({persons.length})
-          </Heading>
-          {persons.length === 0 ? (
-            <Text color="gray.500" textAlign="center" py={8}>
-              No people in this group yet. Add someone above to get started!
-            </Text>
-          ) : (
-            <SimpleGrid columns={{ base: 1, md: 2, lg: 3 }} spacing={4}>
-              {persons.map((person) => (
-                <Card
-                  key={person.id}
-                  p={4}
-                  cursor="pointer"
-                  onClick={() => handlePersonClick(person)}
-                >
-                  <VStack spacing={2}>
-                    <Text fontWeight="bold">{person.name}</Text>
-                    {balances && balances[person.id] && (
-                      <VStack spacing={1}>
-                        <Text fontSize="sm" color="gray.600">
-                          Paid: ${balances[person.id].paid.toFixed(2)}
-                        </Text>
-                        <Text fontSize="sm" color="gray.600">
-                          Owes: ${balances[person.id].owes.toFixed(2)}
-                        </Text>
-                        <Badge
-                          colorScheme={
-                            balances[person.id].balance >= 0 ? "green" : "red"
-                          }
-                        >
-                          {balances[person.id].balance >= 0 ? "+" : ""}$
-                          {balances[person.id].balance.toFixed(2)}
-                        </Badge>
-                      </VStack>
-                    )}
-                    <Button
-                      size="sm"
-                      colorScheme="blue"
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        handleAddExpense(person.id);
-                      }}
-                    >
-                      Add Expense
-                    </Button>
-                  </VStack>
-                </Card>
-              ))}
-            </SimpleGrid>
-          )}
-        </Box>
+        {balances ? (
+          <PersonsList
+            persons={persons}
+            balances={balances}
+            onPersonClick={handlePersonClick}
+            onAddExpense={handleAddExpense}
+          />
+        ) : (
+          <p>Loading balances...</p>
+        )}
 
         {/* Expenses Section */}
-        <Box>
-          <Heading size="lg" mb={4}>
-            Expenses ({expenses.length})
-          </Heading>
-          {expenses.length === 0 ? (
-            <Text color="gray.500" textAlign="center" py={8}>
-              No expenses yet. Click "Add Expense" on a person's card to get
-              started!
-            </Text>
-          ) : (
-            <SimpleGrid columns={{ base: 1, md: 2 }} spacing={4}>
-              {expenses.map((expense) => (
-                <Card key={expense.id} p={4}>
-                  <HStack justify="space-between" align="start">
-                    <VStack align="start" spacing={2} flex={1}>
-                      <Text fontWeight="bold">{expense.name}</Text>
-                      <Text fontSize="lg" color="green.600">
-                        ${expense.amount.toFixed(2)}
-                      </Text>
-                      <Text fontSize="sm" color="gray.600">
-                        Paid by:{" "}
-                        {persons.find((p) => p.id === expense.payer_id)?.name ||
-                          "Unknown"}
-                      </Text>
-                    </VStack>
-                    <IconButton
-                      aria-label="Delete expense"
-                      icon={<DeleteIcon />}
-                      size="sm"
-                      colorScheme="red"
-                      variant="ghost"
-                      onClick={() => handleDeleteExpense(expense.id)}
-                    />
-                  </HStack>
-                </Card>
-              ))}
-            </SimpleGrid>
-          )}
-        </Box>
+        <ExpensesList
+          expenses={expenses}
+          persons={persons}
+          onDeleteExpense={handleDeleteExpense}
+        />
 
         {/* Balances Summary */}
-        {balances && Object.keys(balances).length > 0 && (
-          <Box>
-            <Heading size="lg" mb={4}>
-              Balance Summary
-            </Heading>
-            <SimpleGrid columns={{ base: 1, md: 2, lg: 3 }} spacing={4}>
-              {Object.entries(balances).map(
-                ([personId, personFinancials]: [string, PersonFinancials]) => (
-                  <Card key={personId} p={4}>
-                    <Stat>
-                      <StatLabel>{personFinancials.name}</StatLabel>
-                      <StatNumber
-                        color={
-                          personFinancials.balance >= 0
-                            ? "green.500"
-                            : "red.500"
-                        }
-                      >
-                        {personFinancials.balance >= 0 ? "+" : ""}$
-                        {personFinancials.balance.toFixed(2)}
-                      </StatNumber>
-                      <StatHelpText>
-                        {personFinancials.balance >= 0
-                          ? "Should receive"
-                          : "Should pay"}
-                      </StatHelpText>
-                    </Stat>
-                  </Card>
-                )
-              )}
-            </SimpleGrid>
-          </Box>
+        {balances ? (
+          <BalancesSummary balances={balances} />
+        ) : (
+          <p>Loading balances...</p>
         )}
 
         {/* Back to Home */}
